src/hanerma/memory/manager.py
import uuid
import faiss
import numpy as np
from typing import Dict, Any, List
from hanerma.memory.compression.base_tokenizer import BaseHyperTokenizer

class HCMSManager:
    """
    Hyperfast Compressed Memory Store.
    Manages the hybrid routing between FAISS (Vectors) and Neo4j (GraphRAG).
    """
    def __init__(self, tokenizer: BaseHyperTokenizer, embedding_dim: int = 128):
        self.tokenizer = tokenizer
        self.embedding_dim = embedding_dim
        
        # Initialize FAISS for O(1) semantic similarity search
        self.vector_index = faiss.IndexFlatL2(self.embedding_dim)
        self.memory_map: Dict[int, str] = {} # Maps FAISS index to original compressed text ID
        self.current_idx = 0
        
        # Placeholder for Neo4j Driver (GraphRAG)
        self.graph_db = None 
        
        print(f"[HCMS] Memory Store Online. Dimension: {self.embedding_dim}. Index: FAISS FlatL2.")

    def store_atomic_memory(self, session_id: str, raw_text: str, entity_type: str = "context"):
        """
        Compresses the text and stores it in the optimal backend.
        """
        # 1. Compress via the custom tokenizer adapter
        compressed_tokens = self.tokenizer.encode_and_compress(raw_text)
        efficiency = self.tokenizer.get_compression_ratio(raw_text, compressed_tokens)
        
        print(f"[HCMS] Session {session_id} | Token Compression: -{efficiency}% overhead.")
        
        # 2. Generate a simulated embedding vector for FAISS
        # In production, this calls a lightweight embedding model on the compressed tokens
        vector = np.random.random((1, self.embedding_dim)).astype('float32')
        
        # 3. Store in FAISS
        self.vector_index.add(vector)
        self.memory_map[self.current_idx] = raw_text
        self.current_idx += 1
        
        # 4. If entity_type is specific, route a node to Neo4j
        if entity_type == "fact":
            self._store_in_graph(session_id, raw_text)

    def retrieve_relevant_context(self, query: str, top_k: int = 3) -> List[str]:
        """
        Fetches the most relevant atomic units for the orchestrator to inject into the prompt.
        """
        if self.current_idx == 0:
            return []
            
        # Simulate query embedding
        query_vector = np.random.random((1, self.embedding_dim)).astype('float32')
        
        # Hyperfast FAISS search
        distances, indices = self.vector_index.search(query_vector, top_k)
        
        results = []
        for idx in indices[0]:
            if idx in self.memory_map:
                results.append(self.memory_map[idx])
                
        return results

    def _store_in_graph(self, session_id: str, fact: str):
        """Simulates writing a semantic node to Neo4j for exact relationship tracking."""
        # e.g., graph_db.execute("CREATE (n:Fact {session: $sid, content: $fact})", ...)
        pass
